FixedSizeArrayQueue
Implement the class FixedSizeArrayQueue, which uses a fixed-size array to store items.

Your queue must store any type.

Specification

Constructor: takes in maximum size.
enqueue method: adds to the end of the list (returns void).
Must throw an IllegalArgumentException if a null item is given.
Must throw IllegalStateException if the queue is full.
dequeue method: removes from the front of the list (accepts nothing).
If the queue is empty then this method must throw an IllegalStateException.
peek method: returns, but does not remove, the item at the front of the list (accepts nothing).
If the queue is empty then this method must throw an IllegalStateException.
isEmpty method: tests to see if the queue is empty.
size method: returns the number of items in the queue.
Your implementation must avoid loitering. Loitering is holding a reference to an object when it is no longer needed. The Java garbage collector can reclaim memory for an object only if there are no outstanding references to it. So eliminating these references (by setting them to null) saves memory.